float calc

+++ src/meow_editor/render/editor_window.cpp
@@ -94,8 +94,8 @@ namespace Meow
                     gameobject_ptr->SetName("Sphere");
                     auto transform_ptr = TryAddComponent(
                         gameobject_ptr, "Transform3DComponent", std::make_shared<Transform3DComponent>());
-                    transform_ptr->position = glm::vec3(col * spacing - (float)column_number / 2.0f * spacing,
-                                                        row * spacing - (float)row_number / 2.0f * spacing,
+                    transform_ptr->position = glm::vec3((float)col * spacing - (float)column_number / 2.0f * spacing,
+                                                        (float)row * spacing - (float)row_number / 2.0f * spacing,
                                                         0.0f);
 
                     auto model_comp_ptr =
